feat(client/routes): added private and public stacks
